+++ hc/api/tests/test_notify_mattermost.py
@@ -4,7 +4,7 @@ from datetime import timedelta as td
 from unittest.mock import patch
 
 from django.utils.timezone import now
-from hc.api.models import Channel, Check, Notification
+from hc.api.models import Channel, Check, Notification, Ping
 from hc.test import BaseTestCase
 from django.test.utils import override_settings
 
@@ -18,11 +18,29 @@ class NotifyMattermostTestCase(BaseTestCase):
         self.check.last_ping = now() - td(minutes=61)
         self.check.save()
 
+        self.ping = Ping(owner=self.check)
+        self.ping.created = now() - td(minutes=61)
+        self.ping.save()
+
         self.channel = Channel(project=self.project)
         self.channel.kind = "mattermost"
-        self.channel.value = "123"
+        self.channel.value = "https://example.org"
         self.channel.save()
         self.channel.checks.add(self.check)
+
+    @patch("hc.api.transports.requests.request")
+    def test_it_works(self, mock_post):
+        mock_post.return_value.status_code = 200
+
+        self.channel.notify(self.check)
+        assert Notification.objects.count() == 1
+
+        args, kwargs = mock_post.call_args
+        self.assertEqual(args[1], "https://example.org")
+
+        attachment = kwargs["json"]["attachments"][0]
+        fields = {f["title"]: f["value"] for f in attachment["fields"]}
+        self.assertEqual(fields["Last Ping"], "Success, an hour ago")
 
     @override_settings(MATTERMOST_ENABLED=False)
     def test_it_requires_mattermost_enabled(self):